+++ CHANGES.md
@@ -7,6 +7,8 @@ This version of pyfakefs does not support Python 3.3. Python 3.3 users shall
 keep using pyfakefs 3.4.3, or upgrade to a newer Python version.
 
 #### New Features
+  * a module imported as another name (`import os as _os`) is now correctly 
+    patched without the need of additional parameters
   * added possibility to set root user ([#431](../../issues/431))
   * automatically patch `Path` if imported like `from pathlib import Path` 
     ([#440](../../issues/440))